enable more tests on windows
