+++ pkg/osquery/runtime/runtime_posix_test.go
@@ -21,71 +21,15 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestRestart(t *testing.T) {
-	t.Parallel()
-	runner, teardown := setupOsqueryInstanceForTests(t)
-	defer teardown()
-
-	previousStats := runner.instance.stats
-
-	require.NoError(t, runner.Restart())
-	waitHealthy(t, runner)
-
-	require.NotEmpty(t, runner.instance.stats.StartTime, "start time should be set on latest instance stats after restart")
-	require.NotEmpty(t, runner.instance.stats.ConnectTime, "connect time should be set on latest instance stats after restart")
-
-	require.NotEmpty(t, previousStats.ExitTime, "exit time should be set on last instance stats when restarted")
-	require.NotEmpty(t, previousStats.Error, "stats instance should have an error on restart")
-
-	previousStats = runner.instance.stats
-
-	require.NoError(t, runner.Restart())
-	waitHealthy(t, runner)
-
-	require.NotEmpty(t, runner.instance.stats.StartTime, "start time should be added to latest instance stats after restart")
-	require.NotEmpty(t, runner.instance.stats.ConnectTime, "connect time should be added to latest instance stats after restart")
-
-	require.NotEmpty(t, previousStats.ExitTime, "exit time should be set on instance stats when restarted")
-	require.NotEmpty(t, previousStats.Error, "stats instance should have an error on restart")
+func requirePgidMatch(t *testing.T, pid int) {
+	pgid, err := syscall.Getpgid(pid)
+	require.NoError(t, err)
+	require.Equal(t, pgid, pid)
 }
 
-// TestExtensionIsCleanedUp tests that the osquery extension cleans
-// itself up. Unfortunately, this test has proved very flakey on
-// circle-ci, but just fine on laptops.
-func TestExtensionIsCleanedUp(t *testing.T) {
-	t.Skip("https://github.com/kolide/launcher/issues/478")
-	t.Parallel()
-
-	runner, teardown := setupOsqueryInstanceForTests(t)
-	defer teardown()
-
-	osqueryPID := runner.instance.cmd.Process.Pid
-
-	pgid, err := syscall.Getpgid(osqueryPID)
-	require.NoError(t, err)
-	require.Equal(t, pgid, osqueryPID, "pgid must be set")
-
-	require.NoError(t, err)
-
-	// kill the current osquery process but not the extension
-	err = syscall.Kill(osqueryPID, syscall.SIGKILL)
-	require.NoError(t, err)
-
-	// We need to (a) let the runner restart osquery, and (b) wait for
-	// the extension to die. Both of these may take up to 30s. We'll
-	// start a clock, wait for the respawn, and after 32s, test that the
-	// extension process is no longer running. See
-	// https://github.com/kolide/launcher/pull/342 and associated for
-	// background.
-	timer1 := time.NewTimer(35 * time.Second)
-
-	// Wait for osquery to respawn
-	waitHealthy(t, runner)
-
-	// Ensure we've waited at least 32s
-	<-timer1.C
-}
-
+// TestOsquerySlowStart tests that the launcher can handle a slow-starting osqueryd process.
+// This this is only enabled on non-Windows platforms because suspending we have not yet
+// figured out how to suspend a process on windows via golang.
 func TestOsquerySlowStart(t *testing.T) {
 	t.Parallel()
 	rootDirectory, rmRootDirectory, err := osqueryTempDir()
@@ -131,6 +75,8 @@ func TestOsquerySlowStart(t *testing.T) {
 	require.NoError(t, runner.Shutdown())
 }
 
+// TestExtensionSocketPath tests that the launcher can start osqueryd with a custom extension socket path.
+// This is only run on non-windows platforms because the extension socket path is semi random on windows.
 func TestExtensionSocketPath(t *testing.T) {
 	t.Parallel()
 
@@ -170,41 +116,4 @@ func TestExtensionSocketPath(t *testing.T) {
 	assert.Equal(t, "OK", resp.Status.Message)
 
 	require.NoError(t, runner.Shutdown())
-}
-
-// sets up an osquery instance with a running extension to be used in tests.
-func setupOsqueryInstanceForTests(t *testing.T) (runner *Runner, teardown func()) {
-	rootDirectory, rmRootDirectory, err := osqueryTempDir()
-	require.NoError(t, err)
-
-	k := typesMocks.NewKnapsack(t)
-	k.On("OsqueryHealthcheckStartupDelay").Return(0 * time.Second).Maybe()
-	k.On("WatchdogEnabled").Return(true)
-	k.On("WatchdogMemoryLimitMB").Return(150)
-	k.On("WatchdogUtilizationLimitPercent").Return(20)
-	k.On("WatchdogDelaySec").Return(120)
-	k.On("RegisterChangeObserver", mock.Anything, mock.Anything, mock.Anything, mock.Anything, mock.Anything).Maybe()
-	k.On("Slogger").Return(multislogger.NewNopLogger())
-	k.On("PinnedOsquerydVersion").Return("")
-
-	runner = New(
-		k,
-		WithKnapsack(k),
-		WithRootDirectory(rootDirectory),
-		WithOsquerydBinary(testOsqueryBinaryDirectory),
-	)
-	go runner.Run()
-	waitHealthy(t, runner)
-
-	osqueryPID := runner.instance.cmd.Process.Pid
-
-	pgid, err := syscall.Getpgid(osqueryPID)
-	require.NoError(t, err)
-	require.Equal(t, pgid, osqueryPID)
-
-	teardown = func() {
-		defer rmRootDirectory()
-		require.NoError(t, runner.Shutdown())
-	}
-	return runner, teardown
 }
