+++ pkg/osquery/runtime/runtime_windows_test.go
@@ -1,0 +1,10 @@
+//go:build windows
+// +build windows
+
+package runtime
+
+import (
+	"testing"
+)
+
+func requirePgidMatch(_ *testing.T, _ int) {}
